+++ jlrs/src/memory/target/slot_ref.rs
@@ -21,36 +21,36 @@ impl<'target> LocalSlotRef<'target> {
 /// A reference to a slot in a [`GcFrame`].
 ///
 /// [`GcFrame`]: crate::memory::target::frame::GcFrame
-pub struct StackRef<'target> {
+pub struct StackSlotRef<'target> {
     stack: &'target Stack,
     offset: usize,
 }
 
-/// A reference to a slot in a frame.
-impl<'target> StackRef<'target> {
+impl<'target> StackSlotRef<'target> {
     #[inline]
     pub(crate) fn new(stack: &'target Stack, offset: usize) -> Self {
-        StackRef { stack, offset }
+        StackSlotRef { stack, offset }
     }
 }
 
+/// A reference to a slot in a frame.
 pub unsafe trait SlotRef: private::SlotRefPriv {}
 
 unsafe impl<'target> SlotRef for LocalSlotRef<'target> {}
 
-unsafe impl<'target> SlotRef for StackRef<'target> {}
+unsafe impl<'target> SlotRef for StackSlotRef<'target> {}
 
 mod private {
     use std::{os::raw::c_void, ptr::NonNull};
 
-    use super::{LocalSlotRef, StackRef};
+    use super::{LocalSlotRef, StackSlotRef};
     use crate::private::Private;
 
     pub unsafe trait SlotRefPriv {
         unsafe fn set(&self, data: NonNull<c_void>, _: Private);
     }
 
-    unsafe impl<'target> SlotRefPriv for StackRef<'target> {
+    unsafe impl<'target> SlotRefPriv for StackSlotRef<'target> {
         #[inline]
         unsafe fn set(&self, data: NonNull<c_void>, _: Private) {
             self.stack.set_root(self.offset, data.cast());